refactor: include standard nsci in config/internal domain
