+++ src/environment/index.spec.ts
@@ -5,7 +5,7 @@ import path from "path";
 
 import { expect } from "chai";
 
-import * as RC from "../config/nodesecurerc.js";
+import * as RC from "../config/internal/nsci.js";
 
 import { analyzeEnvironmentContext } from "./index.js";
 
@@ -63,7 +63,7 @@ describe("Environment data collection", () => {
         expect(
           (
             await analyzeEnvironmentContext({
-              ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+              ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
               rootDir: getFixtureFolderPath(fixtureEnvironment.yarn.folderName)
             })
           ).lockFile
@@ -76,7 +76,7 @@ describe("Environment data collection", () => {
       it("should find the shrinkwrap at the given location", async () => {
         expect(
           await analyzeEnvironmentContext({
-            ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+            ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
             rootDir: getFixtureFolderPath(
               fixtureEnvironment.shrinkwrap.folderName
             )
@@ -94,7 +94,7 @@ describe("Environment data collection", () => {
         expect(
           (
             await analyzeEnvironmentContext({
-              ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+              ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
               rootDir: getFixtureFolderPath(
                 fixtureEnvironment.packageLock.folderName
               )
@@ -110,7 +110,7 @@ describe("Environment data collection", () => {
         expect(
           (
             await analyzeEnvironmentContext({
-              ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+              ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
               rootDir: getFixtureFolderPath(
                 fixtureEnvironment.noLockFile.folderName
               )
@@ -127,7 +127,7 @@ describe("Environment data collection", () => {
       it("should keep the package-lock file", async () => {
         expect(
           await analyzeEnvironmentContext({
-            ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+            ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
             rootDir: getFixtureFolderPath(
               fixtureEnvironment.multipleLockFiles.folderName
             )
@@ -147,7 +147,7 @@ describe("Environment data collection", () => {
         it("should fallback to 'NODE' strategy", async () => {
           expect(
             await analyzeEnvironmentContext({
-              ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+              ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
               strategy: "NPM_AUDIT",
               rootDir: getFixtureFolderPath(fixtureEnvironment.yarn.folderName)
             })
@@ -161,7 +161,7 @@ describe("Environment data collection", () => {
 
           expect(
             await analyzeEnvironmentContext({
-              ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+              ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
               strategy: "NPM_AUDIT",
               rootDir: getFixtureFolderPath(
                 fixtureEnvironment.noLockFile.folderName
@@ -183,7 +183,7 @@ describe("Environment data collection", () => {
         const SAME_NODE_STRATEGY = "SECURITY_WG";
         expect(
           await analyzeEnvironmentContext({
-            ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+            ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
             strategy: SAME_NODE_STRATEGY,
             rootDir: getFixtureFolderPath(
               fixtureEnvironment.shrinkwrap.folderName
@@ -201,7 +201,7 @@ describe("Environment data collection", () => {
 
         expect(
           await analyzeEnvironmentContext({
-            ...RC.DEFAULT_RUNTIME_CONFIGURATION,
+            ...RC.DEFAULT_NSCI_RUNTIME_CONFIGURATION,
             strategy: SAME_NONE_STRATEGY,
             rootDir: getFixtureFolderPath(fixtureEnvironment.yarn.folderName)
           })
