+++ src/reporters/post-pipeline/console/outcome.ts
@@ -1,6 +1,6 @@
 import pluralize from "pluralize";
 
-import { Warnings } from "../../../config/nodesecurerc.js";
+import { Warnings } from "../../../config/internal/nsci.js";
 import {
   ConsoleMessage,
   consolePrinter
